+++ tests/utils/test_polismath.py
@@ -2,7 +2,7 @@ import pytest
 from reddwarf.utils import polismath
 from tests.fixtures import polis_convo_data
 from reddwarf.utils import matrix as MatrixUtils
-from reddwarf.polis import PolisClient
+from reddwarf.data_loader import Loader
 
 # TODO: Figure out if in-conv is always equivalent. If so, remove this test.
 @pytest.mark.parametrize("polis_convo_data", ["small-no-meta", "small-with-meta", "medium-with-meta", "medium-no-meta"], indirect=True)
@@ -39,9 +39,8 @@ def test_user_vote_count_sanity_check(polis_convo_data):
     # Convert keys from string to int for later comparison.
     expected_user_vote_counts = {int(k):v for k,v in expected_user_vote_counts.items()}
 
-    client = PolisClient()
-    client.load_data(filepaths=[f"{data_path}/votes.json"])
-    real_vote_matrix = MatrixUtils.generate_raw_matrix(votes=client.data_loader.votes_data)
+    loader = Loader(filepaths=[f"{data_path}/votes.json"])
+    real_vote_matrix = MatrixUtils.generate_raw_matrix(votes=loader.votes_data)
     actual_user_vote_counts = real_vote_matrix.count(axis="columns")
 
     assert expected_user_vote_counts == actual_user_vote_counts.to_dict()